--- pages/_app.js
@@ -1,8 +1,0 @@
-// import '../styles/globals.css'
-import 'tailwindcss/tailwind.css';
-
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />;
-}
-
-export default MyApp;
