+++ pages/_app.jsx
@@ -1,0 +1,12 @@
+/* eslint-disable import/no-extraneous-dependencies */
+/* eslint-disable react/react-in-jsx-scope */
+/* eslint-disable react/prop-types */
+/* eslint-disable react/jsx-props-no-spreading */
+// import '../styles/globals.css'
+import 'tailwindcss/tailwind.css';
+
+function MyApp({ Component, pageProps }) {
+  return <Component {...pageProps} />;
+}
+
+export default MyApp;
