+++ pages/pokemon/[pokemonId].jsx
